Eliminate unit conversion factors
In HMC, mass, energy, time, and temperature do not have physical units. Instead of using constants to enforce a specific interpretation of these variables, we apply them directly in the standard way. This avoids awkward scenarios where, for example, a Switching describing a probability is assumed to have units of Angstrom and a mass of g/mol.